Merged features from dev branch:
- SAM/ASAM/GSAM optimizer support
- EMA (Exponential Moving Average)
- Device management (CUDA/MPS auto-detection)
- AMP (Automatic Mixed Precision) training
- wandb integration for experiment tracking
- Weight scheduling for mixed models
- Gradient accumulation
- uv + scikit-build-core build system

Preserved features from modification branch:
- Modified nucleotide support (use_fp, ECFP embedding)
- ShapeCNN model for SHAPE reactivity prediction
- Learning without Forgetting (LwF)
- Ribonanza/JSON dataset support
- Downsampling for SHAPE data

Key conflict resolutions:
- pyproject.toml: Used dev's scikit-build-core, added rdkit/pandas dependencies
- mix.py/mix_linfold.py: Added use_fp parameter to super().__init__
- shape_nll_loss.py: Integrated both LwF and weight scheduling features
- train.py: Combined dev's optimizer/AMP features with modification's dataset support

+++ mxfold2/fold/zuker_bl.py
@@ -94,7 +94,7 @@ class ZukerFoldBL(AbstractFold):
         return next(self.parameters()).device
 
     def make_param_on_cpu(self, param: dict[str, Any]) -> dict[str, Any]:
-        param_on_cpu = { k: v if k=='cnt' else v.to("cpu") for k, v in param.items() }
+        param_on_cpu = { k: v if k=='cnt' else v.to("cpu").to(torch.float32) for k, v in param.items() }
         param_on_cpu = self.clear_count(param_on_cpu)
         return param_on_cpu
